backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.16,<4.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.1,<2.0
django-cors-headers>=4.7,<5.0
twilio>=9.0,<10.0
openpyxl>=3.1,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
openai>=1.14,<2.0
